Began interrupt routines for buttons
Wrote interrupt routines for PB0 and PB0. They still need debouncing but are otherwise nearly complete. Each one updates the z pointer based on the number of stores that have already been performed (storecount) and stores either  0 of 0xFF at that byte. The button routines also need a way to avoid running if too many button presseshave been entered:
cpi storcount, 16
breq main ;or something like this


I've also added a readcount that we can use to count the number of bytes we have read in for displaying on the LEDs. We will use each byte to define 1 bit of our 4 bit message as per the spec.